Got SD plotting and export to work!
This involved adding a SD timecourse field to CellTrack and the
associated min/max values, getters and setters etc. for TempCellTrack
and CellTrackList.

It also involved/spurred extensive reorganization of the plotting code
in Movie_Explorer_2; the functions drawText and drawCurves were
eliminated and merged into a single drawPlot function. This function is
very long and could certainly be refactored further, but that is for
another day.